+++ cms/sanity.config.ts
@@ -9,7 +9,7 @@ import {
 import {deskStructure} from './structure'
 import {PluginConfig} from './structure/documentInternationalization'
 import {muxInput} from 'sanity-plugin-mux-input'
-import {userGuide} from './structure/CustomTool'
+import {userGuide} from './structure/UserGuide'
 import {media} from 'sanity-plugin-media'
 
 //singleton pages. Before you add the type to singletontypes, the page should be created, since create is not a valid action for singleton types
